multiarch: remove compare_rpm_packages
It is likely that different platforms will use not only
different package sets but even different n-v-rs of the
same packages -- for instance, the kernel.

As a result, the build system should not prevent this
but merely record the versions used. It will be the job
of sanity-checking/compare-to-previous tools to allow
changes to be reviewed/waived.

Signed-off-by: Tim Waugh <twaugh@redhat.com>