users: describe "unique", "primary", and "floating" tags
Rewrite the "Image tags" section to describe the three categories of
tags that OSBS uses: "unique", "primary", and "floating".

Drop the first paragraph from the "Isolated Builds" section since it
duplicates the Image tags section above. Use the term "floating tags"
here instead. This should hint to the user to scroll up and read about
what a "floating tag" is.

Update the final paragraph in the "Isolated Builds" section so that our
definitions of "primary tag" and "unique tag" match what the "Image
tags" section.