Document changes to build_annotations.json
CLOUDBLD-8889

The file is now generated from Tekton results instead, and it is no
longer possible to configure the included annotations via a whitelist.

The task_annotations_whitelist configuration option referenced in the
removed content is not present anywhere else in the docs.

Remove the mention of the remote_source_url annotation entirely, it has
no meaning in the admins doc outside of the whitelisting context. Not to
mention that this annotation is outdated (replaced by remote_sources).

Signed-off-by: Adam Cmiel <acmiel@redhat.com>